#  Makes Tracer.currentSpanCustomizer() lazy 