--- a/brave/src/main/java/brave/LazySpan.java
+++ b/brave/src/main/java/brave/LazySpan.java
@@ -28,7 +28,7 @@ import brave.propagation.TraceContext;
 final class LazySpan extends Span {
   final Tracer tracer;
   final TraceContext context;
-  volatile Span delegate;
+  Span delegate;
 
   LazySpan(Tracer tracer, TraceContext context) {
     this.tracer = tracer;