@lambcode noticed that calling `Tracer.currentSpan()` without using the
result can lead to orphaned spans. The simplest way to prevent this is
to defer overhead until data is added with the result.

Fixes #920

Added benchmarks..

before

```
TracerBenchmarks.currentSpan:currentSpan·p0.999                          sample             8.560             us/op
TracerBenchmarks.currentSpan:·gc.alloc.rate.norm                         sample      15    88.012 ±   0.005    B/op
TracerBenchmarks.currentSpan_tag:currentSpan_tag·p0.999                  sample            20.256             us/op
TracerBenchmarks.currentSpan_tag:·gc.alloc.rate.norm                     sample      15    88.065 ±   0.004    B/op
TracerBenchmarks.currentSpan_unsampled:currentSpan_unsampled·p0.999      sample             4.995             us/op
TracerBenchmarks.currentSpan_unsampled:·gc.alloc.rate.norm               sample      15    24.005 ±   0.001    B/op
```

after

```
TracerBenchmarks.currentSpan:currentSpan·p0.999                          sample              7.974             us/op
TracerBenchmarks.currentSpan:·gc.alloc.rate.norm                         sample      15     24.015 ±   0.008    B/op
TracerBenchmarks.currentSpan_tag:currentSpan_tag·p0.999                  sample             10.192             us/op
TracerBenchmarks.currentSpan_tag:·gc.alloc.rate.norm                     sample      15     72.050 ±  12.511    B/op
TracerBenchmarks.currentSpan_unsampled:currentSpan_unsampled·p0.999      sample              4.198             us/op
TracerBenchmarks.currentSpan_unsampled:·gc.alloc.rate.norm               sample      15     24.009 ±   0.007    B/op
```