@adriancole 2019-06-12
I built this over the commits in #921 to ensure tests would pass considering the use case described