@adriancole 2019-06-12
ps the use cases for calling tracer.currentSpan and not using the result are one of..
* checking null from `tracer.currentSpan()` as a signal for "is current"
* defensively doing some parsing based on whether or not `Span.isNoop()`

just guessing.. I think usually it would be a bug to call `Tracer.currentSpan()` and drop the result.