@adriancole 2019-06-15
I caused a bug https://github.com/apache/incubator-zipkin-brave/pull/927